added test that integrates the Model with the Observer
