+++ tests/Fixtures/SearchableModelWithSensitiveAttributes.php
@@ -3,11 +3,13 @@
 namespace Laravel\Scout\Tests\Fixtures;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
 use Laravel\Scout\Searchable;
 
 class SearchableModelWithSensitiveAttributes extends Model
 {
     use Searchable;
+    use SoftDeletes;
 
     /**
      * The attributes that are mass assignable.
@@ -25,6 +27,13 @@ class SearchableModelWithSensitiveAttributes extends Model
     public function searchShouldUpdate(): bool
     {
         $sensitiveAttributeKeys = ['first_name', 'last_name'];
+
+        $dirtyKeys = collect($this->getDirty())->keys();
+
+        if ($dirtyKeys->isEmpty()) {
+            // Empty dirty attributes means delete/restore.
+            return true;
+        }
 
         return collect($this->getDirty())->keys()
             ->intersect($sensitiveAttributeKeys)
