+++ tests/Unit/ModelObserverTest.php
@@ -4,6 +4,7 @@ namespace Laravel\Scout\Tests\Unit;
 
 use Illuminate\Support\Facades\Config;
 use Laravel\Scout\ModelObserver;
+use Laravel\Scout\Tests\Fixtures\SearchableModelWithSensitiveAttributes;
 use Laravel\Scout\Tests\Fixtures\SearchableModelWithSoftDeletes;
 use Mockery as m;
 use PHPUnit\Framework\TestCase;
@@ -77,6 +78,80 @@ class ModelObserverTest extends TestCase
         $model = m::mock(SearchableModelWithSoftDeletes::class);
         $model->shouldReceive('searchable')->never();
         $model->shouldReceive('unsearchable')->once();
+        $observer->deleted($model);
+    }
+
+    public function test_update_on_sensitive_attributes_triggers_search()
+    {
+        $model = m::mock(
+            new SearchableModelWithSensitiveAttributes([
+                'first_name' => 'taylor',
+                'last_name' => 'Otwell',
+                'remember_token' => 123,
+                'password' => 'secret',
+            ])
+        )->makePartial();
+
+        // Let's pretend it's in sync with the database.
+        $model->syncOriginal();
+
+        // Update
+        $model->password = 'extremelySecurePassword';
+        $model->first_name = 'Taylor';
+
+        // Assertions
+        $model->shouldReceive('searchable')->once();
+        $model->shouldReceive('unsearchable')->never();
+
+        $observer = new ModelObserver;
+        $observer->saved($model);
+    }
+
+    public function test_update_on_non_sensitive_attributes_doesnt_trigger_search()
+    {
+        $model = m::mock(
+            new SearchableModelWithSensitiveAttributes([
+                'first_name' => 'taylor',
+                'last_name' => 'Otwell',
+                'remember_token' => 123,
+                'password' => 'secret',
+            ])
+        )->makePartial();
+
+        // Let's pretend it's in sync with the database.
+        $model->syncOriginal();
+
+        // Update
+        $model->password = 'extremelySecurePassword';
+        $model->remember_token = 456;
+
+        // Assertions
+        $model->shouldReceive('searchable')->never();
+        $model->shouldReceive('unsearchable')->never();
+
+        $observer = new ModelObserver;
+        $observer->saved($model);
+    }
+
+    public function test_unsearchable_should_be_called_when_deleting()
+    {
+        $model = m::mock(
+            new SearchableModelWithSensitiveAttributes([
+                'first_name' => 'taylor',
+                'last_name' => 'Otwell',
+                'remember_token' => 123,
+                'password' => 'secret',
+            ])
+        )->makePartial();
+
+        // Let's pretend it's in sync with the database.
+        $model->syncOriginal();
+
+        // Assertions
+        $model->shouldReceive('searchable')->never();
+        $model->shouldReceive('unsearchable')->once();
+
+        $observer = new ModelObserver;
         $observer->deleted($model);
     }
 }
