+++ tests/Unit/ModelObserverWithSoftDeletesTest.php
@@ -4,6 +4,7 @@ namespace Laravel\Scout\Tests\Unit;
 
 use Illuminate\Support\Facades\Config;
 use Laravel\Scout\ModelObserver;
+use Laravel\Scout\Tests\Fixtures\SearchableModelWithSensitiveAttributes;
 use Laravel\Scout\Tests\Fixtures\SearchableModelWithSoftDeletes;
 use Mockery as m;
 use PHPUnit\Framework\TestCase;
@@ -53,5 +54,27 @@ class ModelObserverWithSoftDeletesTest extends TestCase
         $model->shouldReceive('searchable')->once();
         $model->shouldReceive('unsearchable')->never();
         $observer->restored($model);
+    }
+
+    public function test_unsearchable_should_be_called_when_deleting()
+    {
+        $model = m::mock(
+            new SearchableModelWithSensitiveAttributes([
+                'first_name' => 'taylor',
+                'last_name' => 'Otwell',
+                'remember_token' => 123,
+                'password' => 'secret',
+            ])
+        )->makePartial();
+
+        // Let's pretend it's in sync with the database.
+        $model->syncOriginal();
+
+        // Assertions
+        $model->shouldReceive('searchable')->once();
+        $model->shouldReceive('unsearchable')->never();
+
+        $observer = new ModelObserver;
+        $observer->deleted($model);
     }
 }
